requirements.txt
flask
requests
python-dotenv
groq
flask-cors
pydub
gunicorn